Implementar delegates

+++ src/main/java/com/API/listaTareas/repository/TareasRepository.java
@@ -3,5 +3,5 @@ package com.API.listaTareas.repository;
 import com.API.listaTareas.model.Tarea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TareasRepository extends JpaRepository <Tarea, Long> {
+public interface TareasRepository extends JpaRepository <Tarea, Integer> {
 }
